refactor(balance): extract input indexing
Move the sorted TxIn spending-index rule into Cardano.Node.Client.Inputs and use it from both Balance and TxBuild.
